// fe/src/lib/api/generated/model/appointmentResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentStatus } from './appointmentStatus';
import type { DeliveryMode } from './deliveryMode';
import type { ProviderSummaryResponse } from './providerSummaryResponse';
import type { ServiceSummaryResponse } from './serviceSummaryResponse';

export interface AppointmentResponse {
  cancellation_cutoff_minutes: number | null;
  cancelled_at: string | null;
  created_at: string;
  delivery_mode: DeliveryMode;
  ends_at: string;
  id: string;
  provider: ProviderSummaryResponse;
  reference_code: string;
  service: ServiceSummaryResponse;
  starts_at: string;
  status: AppointmentStatus;
  student_id: string;
}